requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
langchain==0.0.335
requests==2.31.0
tavily-python==0.1.8
pypdf==3.17.1
pydantic==1.10.8
python-dotenv==1.0.0
matplotlib==3.8.2
sqlalchemy==2.0.23
